# docs/swagger.yaml
definitions:
...
  api.VolumeGroup:
    properties:
      name:
        description: Name
        type: string
      nodeNames:
        description: NodeNames
        items:
          type: string
        type: array
      volumeGroupVolumeInfos:
        description: VolumeGroupVolumeInfo
        items:
          $ref: '#/definitions/api.VolumeGroupVolumeInfo'
        type: array
    type: object
  api.VolumeGroupList:
    properties:
      volumeGroupNames:
        description: VolumeGroupNames
        items:
          type: string
        type: array
    type: object
  api.VolumeGroupVolumeInfo:
    properties:
      nodeNames:
        description: NodeNames
        items:
          type: string
        type: array
      state:
        description: State
        type: string
      volumeName:
        description: VolumeName
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /volumegroups/volumegroups:
    get:
      consumes:
      - application/json
      description: get VolumeGroupList
      parameters:
      - description: name
        in: path
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeGroupList'
      summary: 摘要 获取数据卷组列表
      tags:
      - VolumeGroup
  /volumegroups/volumegroups/{name}:
    get:
      consumes:
      - application/json
      description: get VolumeListByVolumeGroup
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeGroup'
      summary: 摘要 获取指定数据卷组中包含数据卷名称列表基本信息
      tags:
      - VolumeGroup
...
swagger: "2.0"